# Docs/openapi_real.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alling B2B API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Listar Pedidos",
        "description": "RF-ORD-001: Lista todos los pedidos del usuario autenticado.\n\nCadena de consulta:\n  1. Obtiene todos los FormatosUnicos del usuario (customer_id = user_id)\n  2. Obtiene todos los Orders relacionados a esos FormatosUnicos\n  → Garantiza aislamiento RLS: el usuario solo ve sus propios pedidos.",
        "operationId": "listar_pedidos_orders__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderSummarySchema"
                  },
                  "type": "array",
                  "title": "Response Listar Pedidos Orders  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/orders/{order_id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Detalle Pedido",
        "description": "RF-ORD-001: Detalle completo de un pedido.\n\nVerifica ownership transitivo:\n  Order.formato_unico_id → FormatoUnico.customer_id == user_id del JWT\nSi la cadena no se cumple → HTTP 403 (no revela existencia del recurso).",
        "operationId": "detalle_pedido_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderDetailSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "formato_unico_id": {
            "type": "string",
            "title": "Formato Unico Id"
          },
          "shipping_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Address"
          },
          "dni_or_ruc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dni Or Ruc"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "shipping_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Cost"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "payment_preference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Preference Id"
          },
          "payment_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Method"
          },
          "cancellation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Reason"
          },
          "cancelled_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled By"
          },
          "order_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Token"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "formato_unico_id",
          "total_amount"
        ],
        "title": "OrderDetailSchema"
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING_PAYMENT",
          "PAID",
          "READY_TO_SHIP",
          "SHIPPED",
          "CANCELLED"
        ],
        "title": "OrderStatus",
        "description": "Estados del FSM de Order (FSM-02).\nTransiciones válidas (MOD-CHK-01):\n  PENDING_PAYMENT → PAID        (ORD-T-02: webhook pago confirmado)\n  PENDING_PAYMENT → CANCELLED   (ORD-T-03: pago fallido / cancelación manual)\n  PAID            → READY_TO_SHIP (ORD-T-04: SELLER confirma despacho)\n  READY_TO_SHIP   → SHIPPED     (ORD-T-05: guía de envío generada)"
      },
      "OrderSummarySchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "shipping_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Cost"
          },
          "payment_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Method"
          },
          "order_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Token"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "formato_unico_id": {
            "type": "string",
            "title": "Formato Unico Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "total_amount",
          "formato_unico_id"
        ],
        "title": "OrderSummarySchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}